change method updateGPA data type, float into double

+++ src/main/java/com/ajou/TwotoSix/unitTest2/repository/MockRepository.java
@@ -9,7 +9,7 @@ public interface MockRepository {
 
     Student findBystudentId(String studentId);
 
-    float updateGPA(Student student, float updatedGPA);
+    double updateGPA(Student student, double updatedGPA);
 
     Student findByStudentId(String studentId);
